# docs/mintlify/openapi.yaml
openapi: 3.0.3
info:
  title: Fuse API
  description: Modular offers platform for Indian e-commerce. Create coupons, auto-applied discounts, EMI offers, and manage checkout.
  version: 1.0.0
  contact:
    name: Fuse
    email: support@fuse.io
    url: https://fuse.io
servers:
  - url: http://localhost:3010
    description: Local development
  - url: https://api.fuse.io
    description: Production
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
      x-default: demo-key-123
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - apiKey: []
paths:
...
  /api/webhooks/{pg_name}:
    post:
      tags: [PG Webhooks]
      summary: PG webhook handler
      description: |
        Receives payment status webhooks from payment gateways.
        No auth required — verified by PG webhook signature.
        Fuse verifies the signature using the merchant's webhook secret,
        then updates the order and session status.
      security: []
      parameters:
        - name: pg_name
          in: path
          required: true
          schema:
            type: string
            description: "razorpay, cashfree, etc."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: PG-specific webhook payload
      responses:
        '200':
          description: Webhook received and processed
        '401':
          description: Webhook signature verification failed

  # === EMI CAMPAIGNS ===
...